Improve install function.
Instead of calling ant on each device to install we now call ant once to
compile the APK file and then use adb to intsall the APK file on each device.
This reduces the installation time on several devices.

As bonus we also check for errors during installation and can warn the user
that the installation failed on a device.