Initial support for building gradle projects.
Android is moving now to use gradle as build platform so I have updated this
plugin so it can handle these projects too.

Gradle support is still partial and has some issues. For example I still cannot
get vim to jump to the correct file when build errors are present.