Remove unused code.
Since this plugin removed support for ant projects we no longer need to get
dependencies from the manifest and since we build the classpath when opening
new java files no need to recurively search for property files withing the
project. When a file of another project is opened the classpath will be updated
with the jars required for that project.