Add basic syntastic support.
This adds basic support for syntastic by setting the
g:syntastic_java_javac_classpath variable with the list of JAR files used in
the android project.

NOTE: I took this from @dambrisco commit but improved it so it does not include
every possible library and version.

NOTE2: I have not tested this personally as I do not use syntastic for
developing.

 [1] - https://github.com/dambrisco/android.vim/commit/154f0468cb26cf774889c4904386c1a23cb77489