Fix #32 - Android detection fails if AndroidManifest.xml has errors.
Implemented a new check to detect if the current file/path is part of an android
project. To do this it searches forward for an AndroidManifest.xml file but
starting from the location of the build.gradle file. This avoids issues like
detecting adjacent android projects:

    projects
      |- MyProject
      |    |- build.gradle
      |
      |- MyAndroidProject
      |    |- build.gradle
      |    |- src
      |        |- main
      |            |- AndroidManifest.xml
      |

If we simply look back and forward for the first AndroidManifest.xml and we are
currently working on a file on MyProject, that is not an android project, then
it will be detected as an android project because it will find the
AndroidManifest.xml on the sibiling project.

By limiting the search to be only forward starting from the build.gradle file
location we ensure it wont find any AndroidManifest.xml within MyProject.