Fix #33 - Vim slowdown due to android#findManifestFile()
Seems that for some projects with large number of files the search for the
AndroidManifest.xml file can take long time. Furthermore this method is called
several times that results in slugish cursor movement.

This commit adds a cache layer to perform the search only once and then use the
cached value on subsequent calls.

TODO: Still need to find a more efficient way to search for the
AndroidManifest.xml file that does not fixes the location as it may change from
project to project.