Improve behavior when Gradle/Android are not found.
When either Gradle or the Android SDK are not found the previous behavior was to
not setup the commands and show error messages when loading java/xml files.

Now if the gradle or the android SDK are not found we set commands that return
error messages with details of why things did not work as expected. The previous
error messages were annoying and disrupted the editor flow.