Add gradle version check.
Sync command require gradle version information to set the gradle
command accordingly to the version, but the gradle version is not
available until Sync command ends!! This created a cyclic dependency.

Not the gradle version is determined when the gradle#version(),
gradle#versionMajor(), or gradle#versionMinor() methods are invoked.
This ensures the version is always available and cached when needed.